+++ docs/src/components/tables/Swatch.tsx
